chore: refactored writeups and media management
