+++ src/app/domains/media/models/media.model.ts
@@ -1,0 +1,10 @@
+export interface MediaMetadata {
+  id: string;
+  fileName: string;
+  mimeType: string;
+  createdAt: Date;
+}
+
+export interface MediaData extends MediaMetadata {
+  data: Uint8Array;
+}
