+++ src/app/domains/writeups/models/writeup.model.ts
@@ -1,0 +1,13 @@
+export interface WriteUpMetadata {
+  id: string;
+  sessionId: string;
+  projectId: string | null;
+  folderId: string | null;
+  name: string;
+  createdAt: Date;
+  updatedAt: Date;
+}
+
+export interface WriteUpData extends WriteUpMetadata {
+  content: string;
+}
